[Core] Move Immutable helper to utils Java package
It is a utility class, but it was packaged with the other utilities.
