+++ aws-datastore/src/main/java/com/amplifyframework/datastore/storage/sqlite/adapter/SQLiteTable.java
@@ -15,13 +15,13 @@
 
 package com.amplifyframework.datastore.storage.sqlite.adapter;
 
-import com.amplifyframework.core.Immutable;
 import com.amplifyframework.core.model.ModelAssociation;
 import com.amplifyframework.core.model.ModelField;
 import com.amplifyframework.core.model.ModelSchema;
 import com.amplifyframework.core.model.types.JavaFieldType;
 import com.amplifyframework.core.model.types.SqliteDataType;
 import com.amplifyframework.core.model.types.internal.TypeConverter;
+import com.amplifyframework.util.Immutable;
 
 import java.util.Collections;
 import java.util.LinkedList;
