+++ core/src/main/java/com/amplifyframework/core/model/ModelSchema.java
@@ -19,13 +19,13 @@ import androidx.annotation.NonNull;
 import androidx.core.util.ObjectsCompat;
 
 import com.amplifyframework.AmplifyException;
-import com.amplifyframework.core.Immutable;
 import com.amplifyframework.core.model.annotations.BelongsTo;
 import com.amplifyframework.core.model.annotations.HasMany;
 import com.amplifyframework.core.model.annotations.HasOne;
 import com.amplifyframework.core.model.annotations.Index;
 import com.amplifyframework.core.model.annotations.ModelConfig;
 import com.amplifyframework.util.FieldFinder;
+import com.amplifyframework.util.Immutable;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
