+++ testmodels/src/main/java/com/amplifyframework/testmodels/commentsblog/AmplifyModelProvider.java
@@ -1,8 +1,8 @@
 package com.amplifyframework.testmodels.commentsblog;
 
-import com.amplifyframework.core.Immutable;
 import com.amplifyframework.core.model.Model;
 import com.amplifyframework.core.model.ModelProvider;
+import com.amplifyframework.util.Immutable;
 
 import java.util.Arrays;
 import java.util.HashSet;
